kafka+clickhouse+postgresql/aggregation/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.31.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
boto3==1.34.69
